1.1.7 (#22)
* Initial push

Initial push

* Cloudflare, fixes in Event

* Changes to event and collaborative log

* Update to datetime to use RFC3339

* Changes to events route

* Changes to event related objects

* FormData Implementation

Event and rewilding image upload. Show created by detail on events and rewilding

* 1.1.0

* 1.1.1

* 1.1.2

- Removal of duplicated reference endpoint in rewilding and event in favour for one general references endpoint
- Changes in datatype for Lat and Lng fields
- Removal of POST rewilding-register

* 1.1.3

- Changes in validation errorlist
- Changes in unauthorized HTTP Code 401 Unauthorised
- Pocket list - Count on move/delete item
- Pocket list - Join on rewilding to get detail
- Pocket List - Location info on rewilding
- Rewilding Search - Water type

* 1.1.4

- Handling of breathing points on participation to event
- Changes to schedule GET and POST structure to be the same
- Changes to Rewilding Create to use same service
- Show participants on user events
- Bugfix pocketlist create rewilding
- Removal of /auth routes as it does not belong to this project

* 1.1.5

- Default type ["national_park", "hiking_area", "campground", "camping_cabin", "park", "playground"] (GET rewilding-search)
- Addition of event_meeting_point_name (POST event | PUT event/{id})
- ⁠Pocket List, Reference link as array (POST rewilding)
- ⁠Adjust event_participants object (GET event | my/event)
- Removal of rewilding_type (GET rewilding)
- Clear unnecessary fmt.Println

* 1.1.6

- Changes to notifications

* 1.1.7

- POST/PUT event missing event type
- Set events_participant_limit & events_payment_fee to 0
- Fix PUT event/{id} response
- POST event/{id}/schedule request and response changes
- GET event/{id}/schedule response changes
- DELETE event/{id}/schedule implement
- DELETE rewilding/{rewildingId} implement

+++ pkg/repository/eventScheduleRepository.go
@@ -7,6 +7,7 @@ import (
 	"oosa_rewild/internal/config"
 	"oosa_rewild/internal/helpers"
 	"oosa_rewild/internal/models"
+	"strconv"
 	"strings"
 	"time"
 
@@ -18,19 +19,12 @@ import (
 type EventScheduleRepository struct{}
 
 type EventScheduleRequest struct {
-	EventId       string                             `json:"event_id" validate:""`
-	EventSchedule []EventScheduleRequestScheduleItem `json:"event_schedule" validate:"required,dive"`
+	EventSchedule [][]EventScheduleRequestTime `json:"schedules" validate:"required,dive,dive"`
 }
 
-type EventScheduleRequestScheduleItem struct {
-	EventSchedulesDate string                     `json:"event_schedules_date" validate:"required,datetime=2006-01-02"`
-	Schedule           []EventScheduleRequestTime `json:"event_schedules_schedules" validate:"dive"`
-	// #TODO: Datetime: RFC-3339 2024
-	// RFC3339     = "2006-01-02T15:04:05Z07:00" 2024-03-24T00:00:00+08:00
-}
 type EventScheduleRequestTime struct {
-	EventSchedulesDatetime    string `json:"event_schedules_datetime" validate:"required,datetime=15:04:05"`
-	EventSchedulesDescription string `json:"event_schedules_description" validate:"required"`
+	EventSchedulesDatetime    string `json:"time" validate:"required,datetime=15:04:05"`
+	EventSchedulesDescription string `json:"description" validate:"required"`
 	// #TODO: Datetime: RFC-3339 2024
 	// RFC3339     = "2006-01-02T15:04:05Z07:00" 2024-03-24T00:00:00+08:00
 }
@@ -44,10 +38,8 @@ type EventScheduleRequestItem struct {
 func (r EventScheduleRepository) Retrieve(c *gin.Context) {
 	eventId := helpers.StringToPrimitiveObjId(c.Param("id"))
 	var Event models.Events
-	Response := EventScheduleRequest{
-		EventId:       c.Param("id"),
-		EventSchedule: []EventScheduleRequestScheduleItem{},
-	}
+
+	Response := [][]EventScheduleRequestTime{}
 	config.DB.Collection("Events").FindOne(context.TODO(), bson.D{{Key: "_id", Value: eventId}}).Decode(&Event)
 
 	var results []models.EventSchedules
@@ -62,7 +54,7 @@ func (r EventScheduleRepository) Retrieve(c *gin.Context) {
 	for _, v := range results {
 		datetime := v.EventSchedulesDatetime.Time()
 		key := datetime.Format("2006-01-02")
-		formattedTime := datetime.Format("15:04:05")
+		formattedTime := datetime.Format("15:04")
 		mappedItem[key] = append(mappedItem[key], EventScheduleRequestTime{
 			EventSchedulesDatetime:    formattedTime,
 			EventSchedulesDescription: v.EventSchedulesDescription,
@@ -77,10 +69,7 @@ func (r EventScheduleRepository) Retrieve(c *gin.Context) {
 			scheduleItem = make([]EventScheduleRequestTime, 0)
 		}
 
-		Response.EventSchedule = append(Response.EventSchedule, EventScheduleRequestScheduleItem{
-			EventSchedulesDate: newDate,
-			Schedule:           scheduleItem,
-		})
+		Response = append(Response, scheduleItem)
 	}
 
 	c.JSON(200, Response)
@@ -142,28 +131,25 @@ func (r EventScheduleRepository) Create(c *gin.Context) {
 	eventId := helpers.StringToPrimitiveObjId(c.Param("id"))
 	eventSchedule := []interface{}{}
 
+	days := int(math.Ceil(Event.EventsDateEnd.Time().Sub(Event.EventsDate.Time()).Hours() / 24))
+
+	if len(payload.EventSchedule) != days {
+		helpers.ResponseBadRequestError(c, "Schedule does not match number of days ("+strconv.Itoa(days)+")")
+		return
+	}
+
 	filters := bson.M{
 		"event_schedules_event": eventId,
 	}
 	config.DB.Collection("EventSchedules").DeleteMany(context.TODO(), filters)
 	config.DB.Collection("Events").FindOne(context.TODO(), bson.D{{Key: "_id", Value: eventId}}).Decode(&Event)
 
-	eventDateStart := Event.EventsDate.Time()
-	eventDateEnd := Event.EventsDateEnd.Time()
+	for i, v := range payload.EventSchedule {
+		scheduleDate := Event.EventsDate.Time().AddDate(0, 0, i).Format("2006-01-02")
 
-	for _, v := range payload.EventSchedule {
-		scheduleDate := v.EventSchedulesDate
-
-		for _, vSchedule := range v.Schedule {
+		for _, vSchedule := range v {
 			fullDatetime := scheduleDate + "T" + vSchedule.EventSchedulesDatetime + "Z"
 			scheduleTime := helpers.StringToDateTime(fullDatetime)
-
-			isValidDate := helpers.TimeIsBetween(scheduleTime, eventDateStart, eventDateEnd)
-
-			if !isValidDate {
-				errList = append(errList, scheduleDate+" "+vSchedule.EventSchedulesDatetime)
-			}
-
 			eventSchedule = append(eventSchedule, models.EventSchedules{
 				EventSchedulesEvent:       helpers.StringToPrimitiveObjId(c.Param("id")),
 				EventSchedulesCreatedBy:   userDetail.UsersId,
@@ -242,6 +228,24 @@ func (r EventScheduleRepository) Delete(c *gin.Context) {
 		config.DB.Collection("EventSchedules").DeleteOne(context.TODO(), filters)
 		helpers.ResultMessageSuccess(c, "Schedule deleted")
 	}
+}
+
+func (r EventScheduleRepository) DeleteAll(c *gin.Context) {
+	err := EventRepository{}.ReadOne(c, &models.Events{})
+	if err != nil {
+		return
+	}
+
+	var Event models.Events
+	eventId := helpers.StringToPrimitiveObjId(c.Param("id"))
+
+	filters := bson.M{
+		"event_schedules_event": eventId,
+	}
+	config.DB.Collection("EventSchedules").DeleteMany(context.TODO(), filters)
+	config.DB.Collection("Events").FindOne(context.TODO(), bson.D{{Key: "_id", Value: eventId}}).Decode(&Event)
+
+	helpers.ResponseSuccessMessage(c, "Event schedule deleted")
 }
 
 func (r EventScheduleRepository) ProcessData(c *gin.Context, EventSchedules *models.EventSchedules, payload EventScheduleRequestItem) {
